PRSD-799: Creates base OccupancyFormPage
